Switch to new video pill placment for beta containers
